backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-requests
python-dotenv
PyJWT
psycopg